research_project/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# File formats
pyyaml>=6.0
bibtexparser>=1.4.0

# Academic APIs
arxiv>=1.4.8
requests>=2.31.0

# NLP / text processing
nltk>=3.8.0
scikit-learn>=1.3.0

# PDF processing
pymupdf>=1.23.0

# Logging
loguru>=0.7.0